% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecat.R
\name{code}
\alias{code}
\title{Define the data coding scheme for a variable}
\usage{
code(variable, levels, labels)
}
\arguments{
\item{variable}{Name of the variable}

\item{levels}{Vector specifying the factor levels}

\item{labels}{Vector specifying the factor labels}
}
\value{
A tibble with three columns and one row per factor level
}
\description{
Define the data coding scheme for a variable
}
\examples{
  code("SEX", levels = c(0, 1), labels = c("Male", "Female"))
}
