% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecat.R
\name{decode}
\alias{decode}
\title{Use codebook to recreate original levels from labelled facors}
\usage{
decode(data, codebook)
}
\arguments{
\item{data}{A data frame or tibble}

\item{codebook}{A codebook}
}
\value{
Data set with variables encoded as factors
}
\description{
Use codebook to recreate original levels from labelled facors
}
\examples{
labels <- codebook(
  code("SEX", levels = c(0, 1), labels = c("Male", "Female")),
  code("STUDY", levels = c(1, 2, 3), labels = c("S1", "S2", "S3"))
)
dataset <- data.frame(
   STUDY = factor(c("S1", "S1", "S1", "S2", "S2", "S2")),
   SEX   = factor(c("Male", "Male", "Female", "Female", "Female", "Male")),
   AGE   = c(32, 18, 64, 52, 26, 80)
)
decode(dataset, labels)
}
