% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecat.R
\name{encode}
\alias{encode}
\title{Apply a codebook to create labelled factors}
\usage{
encode(data, codebook)
}
\arguments{
\item{data}{A data frame or tibble}

\item{codebook}{A codebook}
}
\value{
Data set with variables encoded as factors
}
\description{
Apply a codebook to create labelled factors
}
\examples{
labels <- codebook(
  code("SEX", levels = c(0, 1), labels = c("Male", "Female")),
  code("STUDY", levels = c(1, 2, 3), labels = c("S1", "S2", "S3"))
)
dataset <- data.frame(
   STUDY = c(1, 1, 1, 2, 2, 2),
   SEX   = c(0, 0, 1, 1, 1, 0),
   AGE   = c(32, 18, 64, 52, 26, 80)
)
encode(dataset, labels)
}
